# gin-crud-api/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - full_name
    - password
    - username
    type: object
  domain.UpdateUserRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 100
        minLength: 2
        type: string
      is_active:
        type: boolean
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  domain.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    type: object
  handler.ErrorInfo:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  handler.MetaInfo:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handler.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/handler.ErrorInfo'
      meta:
        $ref: '#/definitions/handler.MetaInfo'
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: CRUD API built with Gin and PostgreSQL.
  title: Gin CRUD API
  version: "1.0"
paths:
  /users:
    get:
      description: Get a paginated list of users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.UserResponse'
                  type: array
                meta:
                  $ref: '#/definitions/handler.MetaInfo'
              type: object
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User creation request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Soft delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Delete user
      tags:
      - users
    get:
      description: Get a user by their UUID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Update user
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"